주석정리,하위뷰 분리
