+++ mohaemukzip/Service/Networking/Auth/LogOutDTO.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct LogoutResponseDTO: Decodable {
+    let isSuccess: Bool
+    let code: String
+    let message: String
+    let result: EmptyResult?
+}
