+++ mohaemukzip/Service/Networking/Auth/ReissueDTO.swift
@@ -1,0 +1,14 @@
+import Foundation
+// MARK: - Auth: Reissue
+
+struct ReissueResponseDTO: Decodable {
+    let isSuccess: Bool
+    let code: String
+    let message: String
+    let result: ReissueResultDTO
+}
+
+struct ReissueResultDTO: Decodable {
+    let accessToken: String
+    let refreshToken: String
+}
