+++ mohaemukzip/Service/Networking/Auth/WithDrawalDTO.swift
@@ -1,0 +1,7 @@
+import Foundation
+struct WithdrawalResponseDTO: Decodable {
+    let isSuccess: Bool
+    let code: String
+    let message: String
+    let result: EmptyResult?
+}
