+++ mohaemukzip/Service/Profile/ProfileEndpoints.swift
@@ -1,10 +1,3 @@
-//
-//  ProfileEndpoints.swift
-//  mohaemukzip
-//
-//  Created by 고석현 on 2/2/26.
-//
-
 import Foundation
 import Moya
 import Alamofire
