+++ mohaemukzip/Sources/Feature/Main/RootView.swift
@@ -1,9 +1,4 @@
-//
-//  RootView.swift
-//  mohaemukzip
-//
-//  Created by 이서현 on 1/30/26.
-//
+
 
 import SwiftUI
 import Combine
