+++ mohaemukzip/Sources/Feature/Profile/Views/ProfileSettingsView.swift
@@ -1,14 +1,4 @@
-//
-//  ProfileSettingsView.swift
-//  mohaemukzip
-//
-//  Created by 고석현 on 2/3/26.
-//
-
 import SwiftUI
-
-
-
 
 struct ProfileSettingsView: View {
