+++ mohaemukzip/Sources/Feature/Recipe/RecipeDetail/RecipeDetailFormatters.swift
@@ -1,9 +1,3 @@
-//
-//  RecipeDetailFOrmatters.swift
-//  mohaemukzip
-//
-//  Created by 고석현 on 2/11/26.
-//
 import SwiftUI
 
 enum RecipeDetailFormatters {
