+++ mohaemukzip/Sources/Feature/Recipe/RecipeDetail/RecipeDetailView.swift
@@ -1,10 +1,3 @@
-//
-//  RecipeDetailView.swift
-//  mohaemukzip
-//
-//  Created by 고석현 on 2/11/26.
-//
-
 import SwiftUI
 
 struct RecipeDetailView: View {
