+++ mohaemukzip/Sources/Feature/Recipe/RecipeDetail/RecipeVideoDetailView.swift
@@ -1,10 +1,3 @@
-//
-//  RecipeVideoDetailView.swift
-//  mohaemukzip
-//
-//  Created by 고석현 on 2/11/26.
-//
-
 import SwiftUI
 import YouTubePlayerKit
 import UIKit
