+++ mohaemukzip/mohaemukzipApp.swift
@@ -1,9 +1,4 @@
-//
-//  mohaemukzipApp.swift
-//  mohaemukzip
-//
-//  Created by 이한결 on 1/8/26.
-//
+
 
 import SwiftUI
